--- a/test/test_ops.py
+++ b/test/test_ops.py
@@ -711,6 +711,8 @@ class TestOps(unittest.TestCase):
     arg = (4,3,2,6)
     helper_test_op([(4,3,1,6)], lambda x: x.expand(arg), lambda x: x.expand(shape=arg))
     helper_test_op([()], lambda x: x.expand([]), lambda x: x.expand(shape=[]))
+    arg = (1,2,3,4,-1,1,6)
+    helper_test_op([(4,3,1,6)], lambda x: x.expand(arg), lambda x: x.expand(shape=arg))
 
   @unittest.skip("very slow")
   def test_sd_big_conv(self):