This basically allows expand to match torch.expand. This PR may not be worth the feature wise. 
Also, this PR adds little bit of performance but not sure if it is real or variance.
`METAL=1 python3.11 examples/llama.py --prompt "Hello." --count 10 --temperature 0 --timing --quantize`

This PR,
```
loaded weights in 32.55 ms, 6.87 GB loaded at 211.15 GB/s
Hello.
ran model in 278.31 ms
sync in 1701.45 ms
 I
ran model in 200.05 ms
sync in 272.64 ms
'
ran model in 164.75 ms
sync in 4.52 ms
m
ran model in 161.49 ms
sync in 4.52 ms
 a
ran model in 165.52 ms
sync in 4.71 ms
 new
ran model in 175.20 ms
sync in 7.14 ms
bie
ran model in 163.20 ms
sync in 4.41 ms
 here
ran model in 165.88 ms
sync in 4.69 ms
.
ran model in 176.89 ms
sync in 4.42 ms
 I
ran model in 164.72 ms
sync in 4.42 ms
```

Master,
```
loaded weights in 34.75 ms, 6.87 GB loaded at 197.81 GB/s
Hello.
ran model in 311.49 ms
sync in 2807.68 ms
 I
ran model in 182.19 ms
sync in 2029.03 ms
'
ran model in 179.55 ms
sync in 7.05 ms
m
ran model in 167.57 ms
sync in 4.53 ms
 a
ran model in 167.82 ms
sync in 4.44 ms
 new
ran model in 179.39 ms
sync in 5.91 ms
bie
ran model in 170.73 ms
sync in 4.62 ms
 here
ran model in 181.93 ms
sync in 4.46 ms
.
ran model in 166.96 ms
sync in 4.67 ms
 I
ran model in 178.66 ms
sync in 4.54 ms
```

Note: Currently it disables `# move permutes before expands (always, this is safe)` in lazy

Note: Based on feedback I can close this PR as if this adds more complexity