@tinyb0t 2023-09-01
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/lazy.py                           3               3
tinygrad/mlops.py                          2               1
tinygrad/shape/shapetracker.py             5               3
tinygrad/tensor.py                         8               5
------------------------------------------------------------
total                                     18              12
------------------------------------------------------------
lines added in the tinygrad folder: 6
```