@ZenginU 2023-09-01
> This should only be in tensor.py. I don't even understand what you mean by torch API.

Hello, what I meant by torch api is the behaviour of torch.extend is not same as Tensor.extend.

`...creates a new view on the existing tensor where a dimension of size one is expanded to a larger size by setting the stride to 0. Any dimension of size 1 can be expanded to an arbitrary value without allocating new memory.`

By editing shapetracker and mlops I tried to mimic the pytorch behaviour.