@geohot 2023-09-01
No. This stuff belongs in tensor.py only